Move routines to types
